+++ examples/workbench/pdf/solutions/no-pdf-skill-needed/solution.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+set -eu
+
+printf 'done\n' > note.txt